Reconfigured urc_navigation package
